chore: move popup code to public directory and change webpack setting
