+++ tsconfig.json
@@ -5,7 +5,7 @@
     "target": "ES5",
     "esModuleInterop": true,
     "sourceMap": true,
-    "rootDir": "src",
+    "rootDir": ".",
     "outDir": "dist/js",
     "noImplicitAny": true,
     "lib": ["dom"]
