+++ webpack.config.js
@@ -6,7 +6,7 @@ const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 module.exports = {
   mode: process.env.ENV || "development",
   entry: {
-    popup: path.resolve(__dirname, "src", "popup.ts"),
+    popup: path.resolve(__dirname, "public", "popup.ts"),
     content: path.resolve(__dirname, "src", "content.ts"),
     background: path.resolve(__dirname, "src", "background.ts"),
   },
